+++ mopidy/backends/dummy.py
@@ -39,7 +39,7 @@ class DummyLibraryProvider(base.BaseLibraryProvider):
         self.dummy_library = []
 
     def find_exact(self, **query):
-        return Playlist()
+        return []
 
     def lookup(self, uri):
         matches = filter(lambda t: uri == t.uri, self.dummy_library)
@@ -50,7 +50,7 @@ class DummyLibraryProvider(base.BaseLibraryProvider):
         pass
 
     def search(self, **query):
-        return Playlist()
+        return []
 
 
 class DummyPlaybackProvider(base.BasePlaybackProvider):