+++ mopidy/backends/local/library.py
@@ -4,7 +4,7 @@ import logging
 
 from mopidy import settings
 from mopidy.backends import base
-from mopidy.models import Playlist, Album
+from mopidy.models import Album
 
 from .translator import parse_mpd_tag_cache
 
@@ -67,7 +67,7 @@ class LocalLibraryProvider(base.BaseLibraryProvider):
                     result_tracks = filter(any_filter, result_tracks)
                 else:
                     raise LookupError('Invalid lookup field: %s' % field)
-        return Playlist(tracks=result_tracks)
+        return result_tracks
 
     def search(self, **query):
         self._validate_query(query)
@@ -101,7 +101,7 @@ class LocalLibraryProvider(base.BaseLibraryProvider):
                     result_tracks = filter(any_filter, result_tracks)
                 else:
                     raise LookupError('Invalid lookup field: %s' % field)
-        return Playlist(tracks=result_tracks)
+        return result_tracks
 
     def _validate_query(self, query):
         for (_, values) in query.iteritems():